# perf: improve type checking performance of BuildMany